+++ doc.go
@@ -1,21 +1,3 @@
-//
-//go:generate ./stamp.sh
-
-// Embed the custom service handlers
-// TODO: Once AWS lambda supports golang as first class, move the
-// NodeJS custom action helpers into golang
-//go:generate go run $GOPATH/src/github.com/mjibson/esc/main.go -o ./CONSTANTS.go -private -pkg sparta ./resources
-//go:generate go run ./resources/awsbinary/insertTags.go ./CONSTANTS !lambdabinary
-
-// Create a secondary CONSTANTS_AWSBINARY.go file with empty content.  The next step will insert the
-// build tags at the head of each file so that they are mutually exclusive, similar to the
-// lambdabinaryshims.go file
-//go:generate go run $GOPATH/src/github.com/mjibson/esc/main.go -o ./CONSTANTS_AWSBINARY.go -private -pkg sparta ./resources/awsbinary/README.md
-//go:generate go run ./resources/awsbinary/insertTags.go ./CONSTANTS_AWSBINARY lambdabinary
-
-// cleanup
-//go:generate rm -f ./resources/provision/node_modules.zip
-
 /*
 Package sparta transforms a set of golang functions into an Amazon Lambda deployable unit.