+++ sparta.go
@@ -17,15 +17,13 @@ import (
 	"strings"
 	"time"
 
-	spartaCF "github.com/mweagle/Sparta/aws/cloudformation"
-	spartaIAM "github.com/mweagle/Sparta/aws/iam"
-
-	"github.com/aws/aws-sdk-go/aws/session"
-	gocf "github.com/mweagle/go-cloudformation"
-
 	"github.com/Sirupsen/logrus"
-	_ "github.com/aws/aws-sdk-go/service/ecr"  // Ref to have Glide include depends
+	"github.com/aws/aws-sdk-go/aws/session"
+	_ "github.com/aws/aws-sdk-go/service/ecr" // Ref to have Glide include depends
+	spartaCF "github.com/mweagle/Sparta/aws/cloudformation"
 	_ "github.com/mweagle/Sparta/aws/dynamodb" // Ref to have Glide include depends
+	spartaIAM "github.com/mweagle/Sparta/aws/iam"
+	gocf "github.com/mweagle/go-cloudformation"
 )
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -34,21 +32,20 @@ import (
 
 const (
 	// SpartaVersion defines the current Sparta release
-	SpartaVersion = "0.13.2"
+	SpartaVersion = "0.20.0"
 	// NodeJSVersion is the Node JS runtime used for the shim layer
-	NodeJSVersion = "nodejs4.3"
+	NodeJSVersion = "nodejs6.10"
 	// PythonVersion is the Python version used for CGO support
 	PythonVersion = "python3.6"
 	// Custom Resource typename used to create new cloudFormationUserDefinedFunctionCustomResource
 	cloudFormationLambda = "Custom::SpartaLambdaCustomResource"
+	// divider length is the length of a divider
+	dividerLength = 40
 )
 
 var (
 	// internal logging header
-	headerDivider = strings.Repeat("=", 40)
-
-	// internal logging subheader
-	subheaderDivider = strings.Repeat("-", 40)
+	headerDivider = strings.Repeat("=", dividerLength)
 )
 
 // AWS Principal ARNs from http://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html
@@ -239,8 +236,8 @@ type ArbitraryJSONObject map[string]interface{}
 // Package private type to deserialize NodeJS proxied
 // Lambda Event and Context information
 type lambdaRequest struct {
-	Event   json.RawMessage `json:"event"`
-	Context LambdaContext   `json:"context"`
+	Event   json.RawMessage
+	Context LambdaContext
 }
 
 // LambdaContext defines the AWS Lambda Context object provided by the AWS Lambda runtime.
@@ -248,14 +245,13 @@ type lambdaRequest struct {
 // for more information on field values.  Note that the golang version doesn't functions
 // defined on the Context object.
 type LambdaContext struct {
-	AWSRequestID       string `json:"awsRequestId"`
-	InvokeID           string `json:"invokeid"`
-	LogGroupName       string `json:"logGroupName"`
-	LogStreamName      string `json:"logStreamName"`
 	FunctionName       string `json:"functionName"`
-	MemoryLimitInMB    string `json:"memoryLimitInMB"`
 	FunctionVersion    string `json:"functionVersion"`
 	InvokedFunctionARN string `json:"invokedFunctionArn"`
+	MemoryLimitInMB    string `json:"memoryLimitInMB"`
+	AWSRequestID       string `json:"awsRequestId"`
+	LogGroupName       string `json:"logGroupName"`
+	LogStreamName      string `json:"logStreamName"`
 }
 
 // LambdaFunction is the golang function signature required to support AWS Lambda execution.
@@ -269,6 +265,9 @@ type LambdaContext struct {
 // Content written to the ResponseWriter will be used as the
 // response/Error value provided to AWS Lambda.
 type LambdaFunction func(*json.RawMessage, *LambdaContext, http.ResponseWriter, *logrus.Logger)
+
+// HTTPLambdaFunction is a more Go-friendly HTTP handler definition
+type HTTPLambdaFunction func(http.ResponseWriter, *http.Request)
 
 // LambdaFunctionOptions defines additional AWS Lambda execution params.  See the
 // AWS Lambda FunctionConfiguration (http://docs.aws.amazon.com/lambda/latest/dg/API_FunctionConfiguration.html)
@@ -657,6 +656,8 @@ func (resourceInfo *customResourceInfo) export(serviceName string,
 type LambdaAWSInfo struct {
 	// pointer to lambda function
 	lambdaFn LambdaFunction
+	// HTTP handler function
+	httpHandler http.Handler
 	// Role name (NOT ARN) to use during AWS Lambda Execution.  See
 	// the FunctionConfiguration (http://docs.aws.amazon.com/lambda/latest/dg/API_FunctionConfiguration.html)
 	// docs for more info.
@@ -694,7 +695,6 @@ func (info *LambdaAWSInfo) lambdaFunctionName() string {
 	if info.cachedLambdaFunctionName != "" {
 		return info.cachedLambdaFunctionName
 	}
-
 	lambdaFuncName := ""
 	if nil != info.Options &&
 		nil != info.Options.SpartaOptions &&
@@ -704,8 +704,10 @@ func (info *LambdaAWSInfo) lambdaFunctionName() string {
 		// Using the default name, let's at least remove the
 		// first prefix, since that's the SCM provider and
 		// doesn't provide a lot of value...
-		lambdaPtr := runtime.FuncForPC(reflect.ValueOf(info.lambdaFn).Pointer())
-		lambdaFuncName = lambdaPtr.Name()
+		if info.lambdaFn != nil {
+			lambdaPtr := runtime.FuncForPC(reflect.ValueOf(info.lambdaFn).Pointer())
+			lambdaFuncName = lambdaPtr.Name()
+		}
 
 		// Split
 		// cwd: /Users/mweagle/Documents/gopath/src/github.com/mweagle/SpartaHelloWorld
@@ -1000,7 +1002,7 @@ func validateSpartaPreconditions(lambdaAWSInfos []*LambdaAWSInfo, logger *logrus
 			logger.WithFields(logrus.Fields{
 				"CollisionCount": eachCount,
 				"Name":           eachLambdaName,
-			}).Error("Detected logically equivalent function associated with multiple structs")
+			}).Error("HandleAWSLambda")
 			errorText = append(errorText, fmt.Sprintf("Multiple definitions of lambda: %s", eachLambdaName))
 		}
 	}
@@ -1069,6 +1071,36 @@ func NewLambda(roleNameOrIAMRoleDefinition interface{},
 	}
 	if lambda.Options.Timeout <= 0 {
 		lambda.Options.Timeout = 3
+	}
+	return lambda
+}
+
+// HandleAWSLambda registers lambdaHandler with the given functionName
+// using the default lambdaFunctionOptions
+func HandleAWSLambda(functionName string,
+	lambdaHandler http.Handler,
+	roleNameOrIAMRoleDefinition interface{}) *LambdaAWSInfo {
+
+	lambda := &LambdaAWSInfo{
+		httpHandler:         lambdaHandler,
+		Options:             defaultLambdaFunctionOptions(),
+		Permissions:         make([]LambdaPermissionExporter, 0),
+		EventSourceMappings: make([]*EventSourceMapping, 0),
+	}
+	if lambda.Options.SpartaOptions == nil {
+		lambda.Options.SpartaOptions = &SpartaOptions{
+			Name: sanitizedName(functionName),
+		}
+	}
+
+	switch v := roleNameOrIAMRoleDefinition.(type) {
+	case string:
+		lambda.RoleName = roleNameOrIAMRoleDefinition.(string)
+	case IAMRoleDefinition:
+		definition := roleNameOrIAMRoleDefinition.(IAMRoleDefinition)
+		lambda.RoleDefinition = &definition
+	default:
+		panic(fmt.Sprintf("Unsupported IAM Role type: %s", v))
 	}
 	return lambda
 }