+++ sparta_xplatbuild.go
@@ -2,8 +2,14 @@
 
 package sparta
 
+import "github.com/Sirupsen/logrus"
+
 // Support Windows development, by only requiring `syscall` in the compiled
 // linux binary.
 func platformKill(parentProcessPID int) {
+	// NOP
+}
+
+func platformLogSysInfo(logger *logrus.Logger) {
 	// NOP
 }